+++ pkg/test_runner/tool/orphan_files.dart
@@ -13,8 +13,7 @@ library;
 import 'dart:io';
 
 import 'package:analyzer/dart/analysis/analysis_context.dart';
-import 'package:analyzer/dart/analysis/context_builder.dart';
-import 'package:analyzer/dart/analysis/context_locator.dart';
+import 'package:analyzer/dart/analysis/analysis_context_collection.dart';
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:test_runner/src/path.dart';
@@ -65,12 +64,18 @@ void _checkTestDirectory(Directory directory) {
 }
 
 void _initAnalysisContext() {
-  var roots = ContextLocator().locateRoots(includedPaths: ['test']);
-  if (roots.length != 1) {
-    throw StateError('Expected to find exactly one context root, got $roots');
+  var collection = AnalysisContextCollection(
+    includedPaths: ['test'],
+  );
+
+  if (collection.contexts.length != 1) {
+    throw StateError(
+      'Expected to find exactly one context root, '
+      'got ${collection.contexts.length}',
+    );
   }
 
-  _analysisContext = ContextBuilder().createContext(contextRoot: roots[0]);
+  _analysisContext = collection.contexts.single;
 }
 
 void _parseReferences(Set<String> importedPaths, String filePath) {